87: Move get_current_user function to users package
While working on separating user-related functionality (#79) we forgot to move `get_current_user` function to the `src/users` package.

So in the scope of this ticket we need to move `get_current_user` to `src/users/utils.py`.
